Moved the deploy content builders into a 'ContentBuilder' class.
